memory_model: Back the shared memory with /dev/shm
In order to obtain a file descriptor reference onto the shared memory
region representing the guest memory, and mmapped to the VMM address
space, we need to create a temporary shared file based on /dev/shm.

The file descriptor pointing to this file will be used as reference
for other processes to access the guest memory, especially when they
will need to read/write from/to virtqueues directly shared with a
guest driver.

This patch needs to be reworked into a more generic patch that would
enable this shared memory backed by a file feature only when vhost
user is needed.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>